+++ tests/post_training/data/ptq_reference_data.yaml
@@ -52,10 +52,6 @@ torchvision/resnet18_backend_OV_QUANTIZER_NNCF:
   metric_value: 0.6946
 torchvision/resnet18_backend_OV_QUANTIZER_AO:
   metric_value: 0.6959
-torchvision/resnet18_backend_X86_QUANTIZER_NNCF:
-  metric_value: 0.6946
-torchvision/resnet18_backend_X86_QUANTIZER_AO:
-  metric_value: 0.6959
 torchvision/mobilenet_v3_small_BC_backend_FP32:
   metric_value: 0.6766
 torchvision/mobilenet_v3_small_BC_backend_OV:
@@ -78,10 +74,6 @@ torchvision/mobilenet_v3_small_BC_backend_OV_QUANTIZER_NNCF:
   metric_value: 0.6679
 torchvision/mobilenet_v3_small_BC_backend_OV_QUANTIZER_AO:
   metric_value: 0.6561
-torchvision/mobilenet_v3_small_BC_backend_X86_QUANTIZER_NNCF:
-  metric_value: 0.0622
-torchvision/mobilenet_v3_small_BC_backend_X86_QUANTIZER_AO:
-  metric_value: 0.0538
 torchvision/vit_b_16_backend_FP32:
   metric_value: 0.8107
 torchvision/vit_b_16_backend_OV:
@@ -102,10 +94,6 @@ torchvision/vit_b_16_backend_OV_QUANTIZER_NNCF:
   metric_value: 0.80922
 torchvision/vit_b_16_backend_OV_QUANTIZER_AO:
   metric_value: 0.2429
-torchvision/vit_b_16_backend_X86_QUANTIZER_NNCF:
-  metric_value: 0.80922
-torchvision/vit_b_16_backend_X86_QUANTIZER_AO:
-  metric_value: 0.1498
 torchvision/swin_v2_s_backend_FP32:
   metric_value: 0.83712
 torchvision/swin_v2_s_backend_OV:
@@ -125,10 +113,6 @@ torchvision/swin_v2_s_backend_CUDA_FX_TORCH:
 torchvision/swin_v2_s_backend_OV_QUANTIZER_NNCF:
   metric_value: 0.8360
 torchvision/swin_v2_s_backend_OV_QUANTIZER_AO:
-  metric_value: 0.8360
-torchvision/swin_v2_s_backend_X86_QUANTIZER_NNCF:
-  metric_value: 0.8360
-torchvision/swin_v2_s_backend_X86_QUANTIZER_AO:
   metric_value: 0.8360
 timm/crossvit_9_240_backend_CUDA_TORCH:
   metric_value: 0.7275
