+++ tests/post_training/pipelines/base.py
@@ -59,8 +59,6 @@ class BackendType(Enum):
     CUDA_FX_TORCH = "CUDA_FX_TORCH"
     OV_QUANTIZER_NNCF = "OV_QUANTIZER_NNCF"
     OV_QUANTIZER_AO = "OV_QUANTIZER_AO"
-    X86_QUANTIZER_NNCF = "X86_QUANTIZER_NNCF"
-    X86_QUANTIZER_AO = "X86_QUANTIZER_AO"
     ONNX = "ONNX"
     OV = "OV"
     OPTIMUM = "OPTIMUM"
@@ -74,8 +72,6 @@ FX_BACKENDS = [
     BackendType.CUDA_FX_TORCH,
     BackendType.OV_QUANTIZER_NNCF,
     BackendType.OV_QUANTIZER_AO,
-    BackendType.X86_QUANTIZER_NNCF,
-    BackendType.X86_QUANTIZER_AO,
 ]
 OV_BACKENDS = [BackendType.OV, BackendType.OPTIMUM]
